requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
oauthlib==3.2.2
pyaes==1.6.1
Pyrogram==2.0.69
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
TgCrypto==1.2.5
tweepy==4.14.0
urllib3==2.0.3
